:sparkles: :fire: :art: Add feels good and indeed and also refactore a lot
