+++ .env-example
@@ -1,0 +1,2 @@
+DISCORD_TOKEN=%discordToken%
+REDIS_URL=redis://redis:6379
