+++ .gitignore
@@ -26,5 +26,3 @@ node_modules
 
 # app config file
 .env
-
-Dockerfile
