+++ Dockerfile
@@ -1,12 +1,14 @@
 FROM node:boron
 
-WORKDIR /usr/src/app
-RUN npm install -g yarn --quiet
+WORKDIR /usr/src/feely
 
-COPY . .
+COPY package.json .
+
+RUN npm install -g yarn --quiet
+RUN yarn install
+
+COPY . /usr/src/feely/
 
 EXPOSE 3000
-
-ENTRYPOINT ["sh", "./entrypoint.sh"]
 
 CMD ["npm", "run", "dev"]
