+++ package.json
@@ -2,7 +2,7 @@
   "name": "feely",
   "version": "0.0.1",
   "description": "SOSH discord bot",
-  "main": "server.js",
+  "main": "app/server.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "dev": "nodemon server.js",
